MODE-2200 Corrected the behavior of the FederatedDocumentStore and how it accesses binary metadata documents.
The fix for this uncovered another issue. By default, ModeShape replaces large string values with binary
values stored in the binary store. But when a connector (in particular the Git connector) builds a node
document with a large string, ModeShape was swapping the large string with a binary value and attempting
to update the reference count. Rather than doing this, ModeShape should not be replacing the connector's
large strings used as property values with binary values. Amended the commit to include a correction to
the `Connetors` class that always passed to the connectors a `DocumentTranslator` instance with a maximum
Long value for the large string size parameter, thereby preventing the connectors from replacing large
strings with binary values.